chore: moving tagging functions to util [TOL-2408]
